Feat: 플레이어 레벨업 구현
플레이어가 일정 이상의 경험치를 얻으면 레벨이 오릅니다.
레벨이 오르면 기초 공격력과 기초 방어력이 일정 수치 오릅니다.

+++ TxtRPG2/SaveData.cs
@@ -17,7 +17,7 @@ namespace TxtRPG2
         public int Level { get; set; }
         public int Hp { get; set; }
         public int Mp { get; set; }
-        public int Atk { get; set; }
+        public float Atk { get; set; }
         public int Def { get; set; }
         public string Job { get; set; }
         public int Gold { get; set; }
